[ALPS05934067] rdump: Update version to MRDUMP11
Update version to MRDUMP11 for extending sctlr_elX

MTK-Commit-Id: c404179948529fee1ec013ff5f6c16baad375027

CR-Id: ALPS05934067
Feature: Memory RAM Dump (MRDUMP)

+++ drivers/misc/mediatek/aee/mrdump/mrdump_arm64.c
@@ -25,5 +25,5 @@ void mrdump_save_control_register(void *creg)
 		      "=r"(cregs->ttbr0_el1), "=r"(cregs->ttbr1_el1),
 		      "=r"(cregs->sp_el[0]), "=r"(cregs->sp_el[1])
 		      : : "memory");
-	cregs->sctlr_el1 = (uint32_t) tmp;
+	cregs->sctlr_el1 = (uint64_t) tmp;
 }
